gutenberg_api_pro/requirements.txt
asgiref==3.8.1
black==25.1.0
cfgv==3.4.0
click==8.2.0
distlib==0.3.9
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
filelock==3.18.0
flake8==7.2.0
identify==2.6.10
inflection==0.5.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
virtualenv==20.31.2
